+++ pytest_unique/unique.py
@@ -3,7 +3,7 @@
 import string
 import sys
 import time
-from datetime import datetime
+from datetime import datetime, timezone
 from functools import partial
 from pathlib import PurePath
 from random import sample
@@ -17,7 +17,7 @@ from pytest_unique.registry import registry_load
 
 def count_factory():
     """Create a counter that increases with each unique call."""
-    start = int(time.mktime(datetime.utcnow().timetuple()))
+    start = int(time.mktime(datetime.now(timezone.utc).timetuple()))
     return memory_count(start)
 
 
@@ -25,11 +25,11 @@ def count_factory():
 class Unique:
     """Generate data using plugins.
 
-    Plugins are read from the `pytest_unique.unique` entrypoints.
+    Plugins are read from the `pytest_unique` entrypoints.
     """
 
     count = field(factory=count_factory)
-    registry = field(factory=partial(registry_load, "unique"))
+    registry = field(factory=partial(registry_load, "pytest_unique"))
 
     def __call__(self, _name, *args, **kwargs):
         """Invoke the unique plugin."""
@@ -39,7 +39,7 @@ class Unique:
     def get_plugin(self, _name, *args, **kwargs):
         """Find plugin in the registry."""
         try:
-            plugins = self.registry["unique"]
+            plugins = self.registry["pytest_unique"]
         except KeyError:
             raise KeyError("No plugins found") from None